+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Problem title and execution constraints are now included when using the "Copy as Markdown" feature.
 - Vscode Elements, codicon.css, codicon.ttf
+- UI adjustments
 
 ## [1.0.2] - 2026-03-03