+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "atcoder-utils" extension will be documented in this 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [1.0.3] - 2026-03-12
 
 ### Added